requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
fabric==2.5.0
idna==2.10
invoke==1.5.0
paramiko==2.7.2
pycparser==2.20
pyindependentreserve==0.2.0
PyNaCl==1.4.0
requests==2.25.1
schedule==0.6.0
six==1.15.0
urllib3==1.26.2
